Continue fixing up the codebase and making refactorings. Changes PLACEHOLDER_GLYPH to 0 (from 1).
