+++ alacritty_terminal/src/cursor.rs
@@ -41,6 +41,7 @@ pub fn get_cursor_glyph(
     let height = metrics.line_height as i32 + i32::from(offset_y);
     let mut width = metrics.average_advance as i32 + i32::from(offset_x);
     let line_width = cmp::max(width * CURSOR_WIDTH_PERCENTAGE / 100, 1);
+
     // Double the cursor width if it's above a double-width glyph
     if is_wide {
         width *= 2;
